Refactor - move code to one place
