+++ src/key-managment/AwsKeyManagement.cs
@@ -28,24 +28,13 @@ namespace Kamus.KeyManagement
 
             var (encryptedData, iv) = RijndaelUtils.Encrypt(dataKey.ToArray(), Encoding.UTF8.GetBytes(data));
 
-            return $"env${encryptedDataKey}${Convert.ToBase64String(iv)}:{Convert.ToBase64String(encryptedData)}";
+            return EnvelopeEncryptionUtils.Wrap(encryptedDataKey, iv, encryptedData);
 
         }
 
         public async Task<string> Decrypt(string encryptedData, string serviceAccountId)
         {
-            var splitted = encryptedData.Split('$');
-            var regex = new Regex("env\\$(?<encryptedDataKey>.*)\\$(?<iv>.*):(?<encryptedData>.*)");
-            var match = regex.Match(encryptedData);
-
-            if (!match.Success)
-            {
-                throw new InvalidOperationException("Invalid encrypted data format");
-            }
-
-            var encryptedDataKey = match.Groups["encryptedDataKey"].Value;
-            var actualEncryptedData = Convert.FromBase64String(match.Groups["encryptedData"].Value);
-            var iv = Convert.FromBase64String(match.Groups["iv"].Value);
+            var (encryptedDataKey, iv, actualEncryptedData) = EnvelopeEncryptionUtils.Unwrap(encryptedData);
 
             var decryptionResult = await mAmazonKeyManagementService.DecryptAsync(new DecryptRequest
             {
