+++ src/key-managment/EnvelopeEncryptionDecorator.cs
@@ -30,24 +30,13 @@ namespace Kamus.KeyManagement
             var dataKey = RijndaelUtils.GenerateKey(256);
             var (encryptedData, iv) = RijndaelUtils.Encrypt(dataKey, Encoding.UTF8.GetBytes(data));
             var encryptedDataKey = await mMasterKeyManagement.Encrypt(Convert.ToBase64String(dataKey), serviceAccountId, createKeyIfMissing);
-            return $"env${encryptedDataKey}${Convert.ToBase64String(iv)}:{Convert.ToBase64String(encryptedData)}";
+            return EnvelopeEncryptionUtils.Wrap(encryptedDataKey, iv, encryptedData);
 
         }
 
         public async Task<string> Decrypt(string encryptedData, string serviceAccountId)
         {
-            var splitted = encryptedData.Split('$');
-            var regex = new Regex("env\\$(?<encryptedDataKey>.*)\\$(?<iv>.*):(?<encryptedData>.*)");
-            var match = regex.Match(encryptedData);
-
-            if (!match.Success)
-            {
-                throw new InvalidOperationException("Invalid encrypted data format");
-            }
-
-            var encryptedDataKey = match.Groups["encryptedDataKey"].Value;
-            var actualEncryptedData = Convert.FromBase64String(match.Groups["encryptedData"].Value);
-            var iv = Convert.FromBase64String(match.Groups["iv"].Value);
+            var (encryptedDataKey, iv, actualEncryptedData) = EnvelopeEncryptionUtils.Unwrap(encryptedData);
 
             mLogger.Information("Encrypted data mactched envelope encryption pattern");
